GCP/requirements.txt
Flask==1.1.4
firebase-admin==5.0.0
google-cloud-firestore==2.1.0

